pyproject.toml
[tool.poetry]
name = "llmops-product"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = ">=0.104.1"
uvicorn = {extras = ["standard"], version = ">=0.24.0"}
python-multipart = ">=0.0.6"
pydantic = ">=2.5.0"
pydantic-settings = ">=2.1.0"
sqlalchemy = ">=2.0.23"
psycopg2-binary = ">=2.9.9"
pgvector = ">=0.2.4"
langchain = ">=0.1.0"
langchain-openai = ">=0.0.2"
langchain-community = ">=0.0.10"
langchain-text-splitters = ">=0.0.1"
langchain-core = ">=0.0.1"
sentence-transformers = ">=2.2.2"
openai = ">=1.6.0"
python-docx = ">=1.1.0"
pypdf = ">=3.17.0"
tiktoken = ">=0.5.1"
python-dotenv = ">=1.0.0"
alembic = ">=1.12.1"
sse-starlette = ">=1.8.2"
numpy = ">=1.24.0"
httpx = ">=0.24.0"
pytz = "^2025.2"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
